Complete overhaul of Database
Changing from PHP to a comepltely new way of communicating with the database.
A server will be run on a computer which will handle all connnection to the database there.
From inside the android application all that will be done is the communication to that server.
This way there is no PHP code involved, everything is handled in pure Java which we are mor familiar with.