Remove server testing projects
The Java Client project that was used to test the server connection was deleted as it is now implemented in the main project.
The Android project that was used to test the communication between the android application and the server was also deleted for the same reasons. The functionality was tested and implemented in the main project.